+++ TACsharp.API/RestAPI/Models/ReqResUser.cs
@@ -1,0 +1,25 @@
+﻿using Newtonsoft.Json;
+
+namespace TACsharp.API.RestAPI.Models
+{
+    /// <summary>
+    /// ReqRes User object model
+    /// </summary>
+    public class ReqResUser
+    {
+        [JsonProperty("id")]
+        public int Id { get; set; }
+
+        [JsonProperty("email")]
+        public string Email { get; set; }
+
+        [JsonProperty("first_name")]
+        public string FirstName { get; set; }
+
+        [JsonProperty("last_name")]
+        public string LastName { get; set; }
+
+        [JsonProperty("avatar")]
+        public string Avatar { get; set; }
+    }
+}
